requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
decorator==4.0.11
Django==1.10.6
django-autoslug==1.9.3
django-crispy-forms==1.6.1
django-extensions==1.7.8
django-livereload-server==0.2.3
django-registration-redux==1.5
djangorestframework==3.6.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
pluggy==0.4.0
public==0.0.0
py==1.4.33
pydot==1.2.3
pyparsing==2.2.0
requests==2.13.0
self==0.0.0
setupfiles==0.0.0
six==1.10.0
tornado==4.4.3
tox==2.7.0
Unidecode==0.4.20
virtualenv==15.1.0
